Cleanup Test and App Config
closes https://github.com/medizininformatik-initiative/torch/issues/629
