+++ src/main/java/de/medizininformatikinitiative/torch/consent/ConsentCodeMapper.java
@@ -6,7 +6,14 @@ import de.medizininformatikinitiative.torch.model.management.TermCode;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
