+++ src/main/java/de/medizininformatikinitiative/torch/service/PatientBatchToCoreBundleWriter.java
@@ -6,6 +6,7 @@ import de.medizininformatikinitiative.torch.model.management.CachelessResourceBu
 import de.medizininformatikinitiative.torch.model.management.ResourceAttribute;
 import de.medizininformatikinitiative.torch.model.management.ResourceBundle;
 import de.medizininformatikinitiative.torch.model.management.ResourceGroup;
+import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
 import java.util.HashSet;
@@ -15,6 +16,7 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
+@Component
 public class PatientBatchToCoreBundleWriter {
 
     private final CompartmentManager compartmentManager;
