+++ src/main/java/de/medizininformatikinitiative/torch/service/StandardAttributeGenerator.java
@@ -7,12 +7,13 @@ import de.medizininformatikinitiative.torch.model.crtdl.AttributeGroup;
 import de.medizininformatikinitiative.torch.model.crtdl.annotated.AnnotatedAttribute;
 import de.medizininformatikinitiative.torch.model.crtdl.annotated.AnnotatedAttributeGroup;
 import de.medizininformatikinitiative.torch.util.CompiledStructureDefinition;
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-
+@Component
 public class StandardAttributeGenerator {
 
     private final StructureDefinitionHandler profileHandler;
